refactor: extract auth routes into routes_auth.py blueprint
Move /login, /logout, /add_user, /reset_user_password, /delete_user,
and /settings/users/<id>/set_password from app.py into a new
routes_auth.py blueprint named "auth".

Also moved to routes_auth.py:
- The `is_admin_required` decorator (other route modules still in
  app.py now import it from here)
- The `@login_manager.user_loader` registration

`login_manager` itself moves to extensions.py alongside `db`, since
it's an extension singleton. `login_manager.login_view` updates from
'login' to 'auth.login' to match the new blueprint-qualified
endpoint name. URL stays /login; only the endpoint name changes.

Template `url_for(...)` calls updated to the new endpoint names:
- base.html: 'logout' -> 'auth.logout'
- settings.html: 'set_user_password' -> 'auth.set_user_password',
  'delete_user' -> 'auth.delete_user',
  'add_user' -> 'auth.add_user'

Drops now-unused imports from app.py: LoginManager, login_user,
logout_user, current_user, UserMixin (already gone in Step 1),
generate_password_hash (already gone), session, abort, wraps,
check_password_hash.
