+++ templates/settings.html
@@ -361,12 +361,12 @@
             <td class="px-4 py-2">{{ 'Yes' if user.is_admin else 'No' }}</td>
             <td class="px-4 py-2">{{ user.last_login_at.strftime('%Y-%m-%d %H:%M:%S') if user.last_login_at else 'Never' }}</td>
             <td class="px-4 py-2 text-right">
-              <form method="POST" action="{{ url_for('set_user_password', user_id=user.id) }}" class="inline">
+              <form method="POST" action="{{ url_for('auth.set_user_password', user_id=user.id) }}" class="inline">
                 <input type="password" name="new_password" placeholder="New Password" class="form-input text-sm w-40" required />
                 <button type="submit" class="bg-indigo-600 text-white px-2 py-1 rounded text-sm hover:bg-indigo-700">Set Password</button>
               </form>
               {% if not user.is_admin %}
-              <form method="POST" action="{{ url_for('delete_user') }}" class="inline" onsubmit="return confirm('Are you sure you want to delete this user?');">
+              <form method="POST" action="{{ url_for('auth.delete_user') }}" class="inline" onsubmit="return confirm('Are you sure you want to delete this user?');">
                 <input type="hidden" name="user_id" value="{{ user.id }}">
                 <button type="submit" class="bg-red-600 hover:bg-red-500 text-white text-xs px-3 py-1 rounded">Delete</button>
               </form>
@@ -381,7 +381,7 @@
       {% endif %}
 
       <h3 class="text-lg font-semibold text-white mt-8 mb-2">Add New User</h3>
-      <form method="POST" action="{{ url_for('add_user') }}" class="space-y-4 max-w-lg">
+      <form method="POST" action="{{ url_for('auth.add_user') }}" class="space-y-4 max-w-lg">
         <input type="hidden" name="action" value="add_user">
         <div>
           <label class="block text-sm text-gray-300">Username</label>
